[wip] need to demonstrate `--passC` usage in README
When cleaning out my local system-wide installation of Notcurses, I had
forgotten to remove `/usr/local/include/notcurses`, which is why I was able to
successfully use only `--passL`. But once I did remove that directory, it's
necessary to point the compiler with `--passC:"-I..."` to my local clone of the
notcurses repo re: the headers.