refactor: use unboxed success values (#4)
* style: add space after `passL:`
* refactor: use unboxed success values

The reason for boxing success values was the idea that a family of types
descended from `ApiSuccess` could, via their names and their fields' names, be
helpful to users of this library re: understanding the nature/purpose of the
values being returned.

At present, it's rather unclear whether that's the direction things are going
with development of the API, and boxing success values in objects only adds
overhead, so for now remove the overhead.

The best approach to take can be reconsidered in the future.